[FEAT] Live render (#1055)
This feature will allow documentation writers to have a live reload of
there changes.

To test this out check out the this branch:
run the following command: `make vendor docker-build` 

Now you can run the application using the normal docker command:

`docker run --rm -v $(pwd):/project -it typo3-docs:local
--config=Documentation`

This will change nothing, if you want to enable watch mode to rerender
on changes you can run:

`docker run --rm -p 1337:1337 -v $(pwd):/project -it typo3-docs:local
--config=Documentation --watch`

This will start a local webserver on port `1337`, this is available the
application will print it to your console.
http://localhost:1337

This local webserver will automatically inject the required javascript
to enable the hot reload. If you want to see it's enabled you can use
the network inspector of your browser to see if there is a request done
which returns a `101` response which means the connection is promoted to
a websocket connection.

Todo:

- [x] update ddev
- [x] update readme to include the example commands